requirements.txt
python-dotenv
llama_index
llama_index.llms.azure_openai
llama_index.embeddings.azure_openai
streamlit
